pyproject.toml
[project]
name = "blog"
version = "2.0.0"
description = "My personal blog"
authors = [
    {name = "Oye Oloyede",email = "oyekunlemac@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdown (>=3.7,<4.0)",
    "pygments (>=2.19.1,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "jinja2 (>=3.1.5,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
build = "blog.build:main"
